Refactor project styles
